LayoutTests/editing/deleting/delete-emoji-8.html

<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable="true">
&#x1F937;&#x1F937;&#x1F3FB;&#x1F937;&zwj;&#x2640;&#xFE0F;&#x1F937;&#x1F3FB;&zwj;&#x2640;&#xFE0F;&#x1F937;&zwj;&#x2642;&#xFE0F;&#x1F937;&#x1F3FB;&zwj;&#x2642;&#xFE0F;&#x1F937;&zwj;&#x2640;&#x1F937;&#x1F3FB;&zwj;&#x2640;&#x1F937;&zwj;&#x2642;&#x1F937;&#x1F3FB;&zwj;&#x2642;&#x1F57A;&#x1F3FB;</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description("This test verifies that deletions are correct over emoji groups and emoji with variations");
var testElement = document.getElementById('test');
getSelection().setBaseAndExtent(testElement.firstChild, testElement.firstChild.length, testElement.firstChild, testElement.firstChild.length);
Markup.dump("test");
while (testElement.firstChild.length) {
    document.execCommand("Delete");
    Markup.dump("test");
}

</script>
</body>
</html>
