LayoutTests/platform/wincairo/editing/deleting/delete-emoji-8-expected.txt

This test verifies that deletions are correct over emoji groups and emoji with variations

Dump of markup 1:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ­Ъци­ЪЈ╗РђЇРЎѓ­ЪЋ║­ЪЈ╗<#selection-caret>
"

Dump of markup 2:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ­Ъци­ЪЈ╗РђЇРЎѓ­ЪЋ║<#selection-caret>"

Dump of markup 3:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ­Ъци­ЪЈ╗РђЇРЎѓ<#selection-caret>"

Dump of markup 4:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ­Ъци­ЪЈ╗РђЇ<#selection-caret>"

Dump of markup 5:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ­Ъци<#selection-caret>"

Dump of markup 6:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇРЎѓ<#selection-caret>"

Dump of markup 7:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ­ЪциРђЇ<#selection-caret>"

Dump of markup 8:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇРЎђ<#selection-caret>"

Dump of markup 9:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци­ЪЈ╗РђЇ<#selection-caret>"

Dump of markup 10:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ­Ъци<#selection-caret>"

Dump of markup 11:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇРЎђ<#selection-caret>"

Dump of markup 12:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ­ЪциРђЇ<#selection-caret>"

Dump of markup 13:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇРЎѓ№ИЈ<#selection-caret>"

Dump of markup 14:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци­ЪЈ╗РђЇ<#selection-caret>"

Dump of markup 15:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ­Ъци<#selection-caret>"

Dump of markup 16:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇРЎѓ№ИЈ<#selection-caret>"

Dump of markup 17:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ­ЪциРђЇ<#selection-caret>"

Dump of markup 18:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇРЎђ№ИЈ<#selection-caret>"

Dump of markup 19:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци­ЪЈ╗РђЇ<#selection-caret>"

Dump of markup 20:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ­Ъци<#selection-caret>"

Dump of markup 21:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇРЎђ№ИЈ<#selection-caret>"

Dump of markup 22:
| "
­Ъци­Ъци­ЪЈ╗­ЪциРђЇ<#selection-caret>"

Dump of markup 23:
| "
­Ъци­Ъци­ЪЈ╗<#selection-caret>"

Dump of markup 24:
| "
­Ъци­Ъци<#selection-caret>"

Dump of markup 25:
| "
­Ъци<#selection-caret>"

Dump of markup 26:
| <br>
